+++ CHANGELOG.md
@@ -33,6 +33,7 @@ Compatibility:
 * Update `String` methods to return `String` instances when called on a subclass (#2453).
 * Update `String#encode` to support the `:fallback` option (#1391).
 * `Module#alias_method` now returns the defined alias as a symbol(#2499, @gogainda).
+* Implement `Symbol#name` (#2453).
 
 Performance: